8791: avoid a circular dependency by moving the FormattedDocketEntry type out of web-client and into shared
