+++ shared/src/business/utilities/getFormattedCaseDetail.ts
@@ -1,6 +1,7 @@
 import {
   CASE_STATUS_TYPES,
   COURT_ISSUED_EVENT_CODES,
+  type DocketEntryRelation,
   OBJECTIONS_OPTIONS_MAP,
   PAYMENT_STATUS,
   STIPULATED_DECISION_EVENT_CODE,
@@ -21,12 +22,21 @@ import { isMiscellaneousDocketEntry } from '@shared/business/utilities/isMiscell
 import { setServiceIndicatorsForPetitionersOnCase } from '@shared/business/utilities/setServiceIndicatorsForPetitionersOnCase';
 import { type ClientApplicationContext } from '@web-client/applicationContext';
 import { type ClientPublicApplicationContext } from '@web-client/applicationContextPublic';
+import { type IconProp } from '@fortawesome/fontawesome-svg-core';
 import { type RawConsolidatedCaseSummary } from '@shared/business/dto/cases/ConsolidatedCaseSummary';
 import { type RawCorrespondence } from '@shared/business/entities/Correspondence';
 import { type RawPractitioner } from '@shared/business/entities/Practitioner';
 import { type ServerApplicationContext } from '@web-api/applicationContext';
 import { type TAssociatedCase } from '@shared/business/useCases/getCasesForUserInteractor';
 import { type UnknownAuthUser } from '@shared/business/entities/authUser/AuthUser';
+
+export type RelatedDocketEntry = DocketEntryRelation & {
+  dispositionLinkText: string[];
+  dispositionText: [];
+  docketEntryIndex?: number;
+  showDocumentViewerLink: boolean;
+  showDownloadLink: boolean;
+};
 
 export type FormattedCaseInventoryReportEntry = {
   docketNumber: string;
@@ -70,6 +80,30 @@ export type FormattedCaseDetailDocketEntry = RawDocketEntry & {
   signedAtFormatted: string;
   signedAtFormattedTZ: string;
   sortingFilingDate: string;
+};
+
+export type FormattedDocketEntry = FormattedCaseDetailDocketEntry & {
+  editDocketEntryMetaLink: string;
+  iconsToDisplay: {
+    className: string;
+    icon: IconProp;
+    size: string;
+    title: string;
+  }[];
+  relatedDocketEntries: RelatedDocketEntry[];
+  sealButtonText: string;
+  sealButtonTooltip: string;
+  sealIcon: string;
+  showDocumentDescriptionWithoutLink: boolean;
+  showDocumentProcessing: boolean;
+  showDocumentViewerLink: boolean;
+  showEditDocketRecordEntry: boolean;
+  showLinkToDocument: boolean;
+  showLoadingIcon: boolean;
+  showNotServed: boolean;
+  showSealDocketRecordEntry: boolean;
+  showServed: boolean;
+  toolTipText: string;
 };
 
 type FormattedCounsel = RawPractitioner & {
